WINDUP-3720 Decoupled JSON extract generation from reports rendering
