+++ rules-java/api/src/main/java/org/jboss/windup/rules/apps/java/reporting/rules/CreateIssueSummaryDataRuleProvider.java
@@ -1,50 +1,48 @@
 package org.jboss.windup.rules.apps.java.reporting.rules;
 
-import java.io.File;
 import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Path;
-import java.text.SimpleDateFormat;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.fasterxml.jackson.annotation.JsonFilter;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.ser.FilterProvider;
-import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
 import org.jboss.forge.furnace.util.OperatingSystemUtils;
 import org.jboss.windup.config.AbstractRuleProvider;
 import org.jboss.windup.config.GraphRewrite;
 import org.jboss.windup.config.loader.RuleLoaderContext;
+import org.jboss.windup.config.metadata.RuleMetadata;
 import org.jboss.windup.config.operation.GraphOperation;
 import org.jboss.windup.config.phase.ReportRenderingPhase;
-import org.jboss.windup.graph.GraphContext;
 import org.jboss.windup.graph.model.ProjectModel;
 import org.jboss.windup.graph.model.WindupConfigurationModel;
 import org.jboss.windup.graph.model.resource.FileModel;
-import org.jboss.windup.graph.service.GraphService;
 import org.jboss.windup.graph.service.WindupConfigurationService;
 import org.jboss.windup.graph.traversal.OnlyOnceTraversalStrategy;
 import org.jboss.windup.graph.traversal.ProjectModelTraversal;
-import org.jboss.windup.reporting.freemarker.problemsummary.ProblemSummary;
-import org.jboss.windup.reporting.freemarker.problemsummary.ProblemSummaryService;
-import org.jboss.windup.reporting.model.TechnologyUsageStatisticsModel;
-import org.jboss.windup.reporting.service.EffortReportService;
-import org.jboss.windup.reporting.service.ReportService;
 import org.jboss.windup.reporting.category.IssueCategory;
 import org.jboss.windup.reporting.category.IssueCategoryRegistry;
+import org.jboss.windup.reporting.freemarker.problemsummary.ProblemSummary;
+import org.jboss.windup.reporting.freemarker.problemsummary.ProblemSummaryService;
+import org.jboss.windup.reporting.service.EffortReportService;
+import org.jboss.windup.reporting.service.EffortReportService.EffortLevel;
+import org.jboss.windup.reporting.service.ReportService;
 import org.jboss.windup.util.exception.WindupException;
 import org.ocpsoft.rewrite.config.Configuration;
 import org.ocpsoft.rewrite.config.ConfigurationBuilder;
 import org.ocpsoft.rewrite.context.EvaluationContext;
 
+import com.fasterxml.jackson.annotation.JsonFilter;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.MappingJsonFactory;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.jboss.windup.config.metadata.RuleMetadata;
-import org.jboss.windup.reporting.service.EffortReportService.EffortLevel;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.ser.FilterProvider;
+import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
+import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
 
 /**
  * Generates a .js (javascript) file in the reports directory with detailed issue summary information.
@@ -80,7 +78,6 @@ public class CreateIssueSummaryDataRuleProvider extends AbstractRuleProvider {
 
                 issueSummaryWriter.write("var WINDUP_ISSUE_SUMMARIES = [];" + NEWLINE);
                 issueSummaryWriter.write("var WINDUP_TECHNOLOGIES = [];" + NEWLINE);
-                List<Object> analysisSummaryList = new ArrayList<>();
                 for (FileModel inputApplicationFile : windupConfiguration.getInputPaths()) {
                     ProjectModel inputApplication = inputApplicationFile.getProjectModel();
                     ProjectModelTraversal projectModelTraversal = new ProjectModelTraversal(inputApplication, new OnlyOnceTraversalStrategy());
@@ -103,14 +100,6 @@ public class CreateIssueSummaryDataRuleProvider extends AbstractRuleProvider {
                     issueSummaryWriter.write("WINDUP_ISSUE_SUMMARIES['" + inputApplication.getId() + "'] = ");
                     objectMapper.writer(filters).writeValue(issueSummaryWriter, summariesBySeverity);
                     issueSummaryWriter.write(";" + NEWLINE);
-                    if (windupConfiguration.isExportingSummary()) {
-                        analysisSummaryList.add(writeApplicationExportSummary(summariesBySeverity, inputApplicationFile, event.getGraphContext()));
-                    }
-                }
-
-                if (windupConfiguration.isExportingSummary())
-                {
-                    writeJsonOutputFile(analysisSummaryList, windupConfiguration);
                 }
 
                 issueSummaryWriter.write("var effortToDescription = [];" + NEWLINE);
@@ -148,114 +137,5 @@ public class CreateIssueSummaryDataRuleProvider extends AbstractRuleProvider {
     @JsonFilter("graphFilter")
     public static class PropertyFilterMixin {
 
-    }
-
-    private Map<String, Object> writeApplicationExportSummary(Map<String, List<ProblemSummary>> summariesBySeverity, FileModel application, GraphContext context) {
-
-        Map<String, Map<String, Integer>> translatedResults = new HashMap<>();
-
-        summariesBySeverity.forEach((k, v) -> {
-            int incidents = 0;
-            int effortPoints = 0;
-            for (ProblemSummary summary : v) {
-                incidents += summary.getNumberFound();
-                effortPoints += summary.getNumberFound() * summary.getEffortPerIncident();
-            }
-            HashMap<String, Integer> results = new HashMap<>();
-            results.put("incidents", incidents);
-            results.put("totalStoryPoints", effortPoints);
-            translatedResults.put(k, results);
-        });
-
-
-        Map<String, Object> resultsWithTitle = new LinkedHashMap<>();
-        resultsWithTitle.put("application", application.getFileName());
-        resultsWithTitle.put("incidentsByCategory", translatedResults);
-
-        List<ProblemSummary> mandatorySummaries = summariesBySeverity.get("mandatory");
-        Map<Integer, Integer> incidentsByEffort = new HashMap<>();
-        if (mandatorySummaries != null) {
-            mandatorySummaries.forEach(ps -> {
-                if (!incidentsByEffort.containsKey(ps.getEffortPerIncident())) {
-                    incidentsByEffort.put(ps.getEffortPerIncident(), ps.getNumberFound());
-                } else {
-                    incidentsByEffort.replace(ps.getEffortPerIncident(), incidentsByEffort.get(ps.getEffortPerIncident()) + ps.getNumberFound());
-                }
-            });
-        }
-        Map<String, Map<String, Integer>> translatedEffortResults = new HashMap<>();
-
-        incidentsByEffort.forEach((k, v) -> {
-            HashMap<String, Integer> results = new HashMap<>();
-
-            results.put("incidents", v);
-            results.put("totalStoryPoints", k * v);
-            translatedEffortResults.put(getEffortDescription(k), results);
-        });
-
-        resultsWithTitle.put("mandatoryIncidentsByType", translatedEffortResults);
-        resultsWithTitle.put("technologyTags", getTechnologyTagsForApplication(application, context));
-        return resultsWithTitle;
-    }
-
-    private void writeJsonOutputFile(List analysisSummary, WindupConfigurationModel windupConfig){
-        String outputFolderPath = windupConfig.getOutputPath().getFilePath() + File.separator;
-
-        ObjectMapper mapper = new ObjectMapper();
-        String json;
-
-        FileWriter writer = null;
-        try {
-            json = mapper.writeValueAsString(analysisSummary);
-            String filename = "analysisSummary.json";
-            writer = new FileWriter(outputFolderPath + filename);
-            writer.write(json);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Couldn't convert given map to JSON: " + e.getMessage());
-        } catch (IOException ioe) {
-            throw new RuntimeException("Couldn't write summary to file: " + ioe.getMessage());
-        } finally {
-            try {
-                writer.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-    }
-
-    private Set<Map<String, String>> getTechnologyTagsForApplication(FileModel application, GraphContext context) {
-        final GraphService<TechnologyUsageStatisticsModel> service = new GraphService<>(context, TechnologyUsageStatisticsModel.class);
-        return service.findAll()
-                .stream()
-                .filter(technologyUsageStatisticsModel -> application.getProjectModel().equals(technologyUsageStatisticsModel.getProjectModel().getRootProjectModel()))
-                .map(technologyUsageStatisticsModel -> {
-                    final Map<String, String> techTag = new HashMap<>(2);
-                    techTag.put("name", technologyUsageStatisticsModel.getName());
-                    techTag.put("category", technologyUsageStatisticsModel.getTags()
-                                                .stream()
-                                                .filter(technologyName -> !DISCARDED_TAGS.contains(technologyName))
-                                                .findFirst()
-                                                .orElse("Other"));
-                    return techTag;
-                })
-                .collect(Collectors.toCollection(() -> new TreeSet<>(Comparator.comparing(stringStringMap -> stringStringMap.get("name")))));
-    }
-
-    private String getEffortDescription(Integer effort){
-        switch (effort) {
-            case 0:
-                return "Info";
-            case 1:
-                return "Trivial";
-            case 3:
-                return "Complex";
-            case 5:
-                return "Redesign";
-            case 7:
-                return "Architectural";
-            default:
-                return "Unknown";
-        }
     }
 }
